+++ src/components/buttons/VcsToolButton.vue
@@ -2,11 +2,11 @@
   <div class="vcs-button-wrap">
     <v-btn
       v-if="!tooltip"
-      text
       :color="appliedColor"
-      :disabled="disabled"
-      class="vcs-button pa-0"
-      :ripple="false"
+      :text="isTextButton"
+      class="vcs-tool-button"
+      :class="classes"
+      :ripple="{ class: 'primary--text text--darken-4' }"
       elevation="0"
       v-bind="{ ...$attrs }"
       v-on="{ ...$listeners }"
@@ -24,11 +24,11 @@
     >
       <template #activator="{ on, attrs }">
         <v-btn
-          text
           :color="appliedColor"
-          :disabled="disabled"
-          class="vcs-button pa-0"
-          :ripple="false"
+          :text="isTextButton"
+          class="vcs-tool-button"
+          :class="classes"
+          :ripple="{ class: 'primary--text text--darken-4' }"
           elevation="0"
           v-bind="{ ...$attrs, ...attrs }"
           v-on="{ ...$listeners, ...on }"
@@ -54,20 +54,19 @@
     right: -3px;
   }
   .v-btn {
-    &.vcs-button {
-      height: 16px;
-      min-width: auto;
-      &:hover {
-        color: var(--v-primary-lighten1) !important;
+    &.vcs-tool-button {
+      min-width: 42px;
+      height: 34px;
+      padding: 0;
+      position: relative;
+
+      &.vcs-tool-button--active-background {
+        border: 2px solid var(--v-primary-base);
+        background-color: var(--v-base-lighten2) !important;
       }
-      ::v-deep {
-        .v-icon {
-          font-size: 16px;
-        }
-        .v-icon__component {
-          height: 16px;
-          width: 16px;
-        }
+
+      &.v-btn--disabled {
+        background-color: transparent !important;
       }
     }
   }
@@ -79,21 +78,23 @@
   import VcsTooltip from '../notification/VcsTooltip.vue';
 
   /**
-   * @description a button with tooltip extending {@link https://vuetifyjs.com/en/api/v-btn/|vuetify v-btn} using {@link VcsTooltip}.
-   * @vue-prop {boolean}                                active - State of the button. Applies set color or vuetify primary color in active state.
+   * @description a button with tooltip extending {@link https://vuetifyjs.com/en/api/v-btn/|vuetify v-btn} using {@link VcsTooltip}. Used for tool buttons in the Navbar.
+   * @vue-prop {boolean}                                active - Whether button has background color. Applies vuetify primary color if color property is not set.
    * @vue-prop {string}                                 color - Passes property to v-btn in case prop active is true.
-   * @vue-prop {boolean}                                disabled
+   * @vue-prop {boolean}                                background - When applied with active the button shows an active-background state implying that a tool is active, but running in background, e.g. windowComponent closed.
    * @vue-prop {boolean}                                hasUpdate - Whether the button shows a badge in the top right.
    * @vue-prop {string}                                 icon - When given, will display an icon in the button. Replaces vuetify icon property.
    * @vue-prop {string}                                 tooltip - Text content of a tooltip which appears on hover with default delay.
    * @vue-prop {('bottom' | 'left' | 'top' | 'right')}  tooltipPosition - Position of the tooltip.
    * @vue-prop {Object<string, any>}                    tooltipProps - Properties to be passed to VcsTooltip {@link https://vuetifyjs.com/en/api/v-tooltip/#props|vuetify v-tooltip}
-   * @vue-computed {string}                             appliedColor - color applied to button, depending on active state
+   * @vue-computed {string}                             appliedColor - color applied to button, depending on size and state
+   * @vue-computed {Object<string, string>}             classes - css classes applied to button, depending on size and state
    * @vue-computed {boolean}                            hasDefaultSlot
+   * @vue-computed {boolean}                            isTextButton - button is text
    * @vue-event {MouseEvent}                            click - Emits click event when the button is clicked.
    */
   export default {
-    name: 'VcsButton',
+    name: 'VcsToolButton',
     components: {
       VcsTooltip,
       VcsBadge,
@@ -110,7 +111,7 @@
         type: String,
         default: undefined,
       },
-      disabled: {
+      background: {
         type: Boolean,
         default: false,
       },
@@ -143,8 +144,17 @@
           return null;
         }
       },
+      classes() {
+        return {
+          'vcs-tool-button--active-background':
+            this.active && this.background && !this.$attrs.disabled,
+        };
+      },
       hasDefaultSlot() {
         return !!this.$slots?.default?.[0]?.text?.trim();
+      },
+      isTextButton() {
+        return !this.active ? true : null;
       },
     },
   };