+++ src/search/resultItem.vue
@@ -13,7 +13,6 @@
       :actions="item.actions"
       :block-overflow="true"
       :overflow-count="2"
-      small
       right
     />
   </div>